tests/functional/requirements.txt
pytest==6.2.5
pytest-asyncio==0.12.0
aiohttp==3.8.0
elasticsearch[async]==7.9.1
redis==4.5.0
pydantic==1.10.7
orjson
python-dotenv==1.0.0
backoff==2.2.1